# openapi.yml
openapi: "3.0.0"
info:
  title: Simple Plan API
  version: 1.0.0
paths:
...
  /taskLists:
    get:
      x-controller: taskLists
      x-operation: getTaskLists
      tags:
        - Task Lists
      security:
        - cookieAuth: []
      summary: Get task lists.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskLists'
        400:
          description: Invalid request.
        401:
          description: Not authenticated.
    post:
      x-controller: taskLists
      x-operation: createTaskList
      tags:
        - Task Lists
      security:
        - cookieAuth: []
      summary: Add a task list.
      requestBody:
        $ref: '#/components/requestBodies/SetTaskList'
      responses:
        201:
          description: Task list created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        400:
          description: Invalid request.
        401:
          description: Not authenticated.
        409:
          description: Task list already exists.
...
  /tasks:
    get:
      x-controller: tasks
      x-operation: getTasks
      tags:
        - Tasks
      security:
        - cookieAuth: []
      summary: Get tasks.
      parameters:
        - name: completed
          in: query
          schema:
            type: boolean
        - name: duration
          in: query
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tasks'
        400:
          description: Invalid request.
        401:
          description: Not authenticated.
...
components:
...
  requestBodies:
    SetTaskList:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TaskList'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: simplePlanSessionId
  schemas:
...
    TaskList:
      type: object
      required:
        - title
        - completed
      properties:
        listId:
          type: string
          readOnly: true
        title:
          type: string
        completed:
          type: string
          format: date-time
          nullable: true
    TaskLists:
      type: array
      items:
        $ref: '#/components/schemas/TaskList'
    Task:
      type: object
      required:
        - title
        - duration
      properties:
        taskId:
          type: string
          readOnly: true
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        complete:
          type: string
          format: date-time
          nullable: true
    Tasks:
      type: array
      items:
        $ref: '#/components/schemas/Task'
...